+++ app/src/main/java/woowacourse/movie/view/bindingadapter/ImageBindingResource.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie.view
+package woowacourse.movie.view.bindingadapter
 
 import android.widget.ImageView
 import androidx.annotation.DrawableRes
